back-end/requirements.txt
flask
flask_sqlalchemy
flask_jwt_extended
flask_bcrypt
psycopg2
flask_cors
python-dotenv
redis>=5.0
psycopg2-binary
gunicorn